Add limit-width option
When working on very narrow windows I noticed that if the buffer name
was too long, it would cause a floating window that would overflow out
of the pane it was representing. This may be desirable for some, but I'd
rather limit it to the width of the buffer.

That is what the new `limit-width` option provides; when enabled, should
the buffer text not fit in the available width, it will truncate it to
do so.

I can see future expansions to this if you think they are appropriate:

1. don't truncate the text, but have it wrap over multiple lines
2. allow the user to customise more how the buffer name is truncated -
   custom replacements, perhaps?

Please let me know what you think, and if this is something you're happy
to land. Thanks!